Updating viewport
It is now updating the viewport of the portal based on the player, however it is not update with the correct angle. Neither is it rendering the correct things as it is not rendering only from the portal.
It is also not updating based on the "lokal" position of the headset, but from a more general position of the headset "area"